app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "openmeteo-requests>=1.3.0",
    "pandas>=2.2.3",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "streamlit>=1.39.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]
